centralized exception management (unfinish)

+++ ShopApp_Backend/Shopapp/src/main/java/com/project/Shopapp/services/loaisanpham/LoaiSanPhamService.java
@@ -2,9 +2,13 @@ package com.project.Shopapp.services.loaisanpham;
 
 import com.project.Shopapp.dtos.LoaiSanPhamDTO;
 import com.project.Shopapp.models.LoaiSanPham;
+import com.project.Shopapp.models.SanPham;
 import com.project.Shopapp.repositories.LoaiSanPhamRepository;
+import com.project.Shopapp.repositories.SanPhamRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -12,6 +16,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class LoaiSanPhamService implements ILoaiSanPhamService {
     private final LoaiSanPhamRepository loaiSanPhamRepository;
+    private final SanPhamRepository sanPhamRepository;
 
     @Override
     public LoaiSanPham createLoaiSanPham(LoaiSanPhamDTO loaiSanPhamDTO) {
@@ -38,7 +43,17 @@ public class LoaiSanPhamService implements ILoaiSanPhamService {
     }
 
     @Override
-    public void deleteLoaiSanPham(int id) {
-        loaiSanPhamRepository.deleteById(id);
+    @Transactional
+    public LoaiSanPham deleteLoaiSanPham(int id) throws Exception {
+        LoaiSanPham loaiSanPham = loaiSanPhamRepository.findById(id)
+                .orElseThrow(() -> new ChangeSetPersister.NotFoundException());
+
+        List<SanPham> sanPhams = sanPhamRepository.findByMALOAISANPHAM(loaiSanPham);
+        if (!sanPhams.isEmpty()) {
+            throw new IllegalStateException("Cannot delete LoaiSanPham with associated products");
+        } else {
+            loaiSanPhamRepository.deleteById(id);
+            return loaiSanPham;
+        }
     }
 }
